refactor: drop brokerage simple shim
