+++ qmtl/brokerage/__init__.py
@@ -4,14 +4,23 @@
 
 from .order import Order, Fill, Account, OrderType, TimeInForce
 from .interfaces import BuyingPowerModel, FeeModel, SlippageModel, FillModel
-from .simple import (
-    CashBuyingPowerModel,
+from .buying_power import CashBuyingPowerModel, CashWithSettlementBuyingPowerModel
+from .fill_models import (
     ImmediateFillModel,
-    CashWithSettlementBuyingPowerModel,
+    BaseFillModel,
+    MarketFillModel,
+    LimitFillModel,
+    StopMarketFillModel,
+    StopLimitFillModel,
+    UnifiedFillModel,
 )
 from .fees import PerShareFeeModel, PercentFeeModel, CompositeFeeModel, IBKRFeeModel
-from .slippage import NullSlippageModel, ConstantSlippageModel, SpreadBasedSlippageModel, VolumeShareSlippageModel
-from .fill_models import BaseFillModel, MarketFillModel, LimitFillModel, StopMarketFillModel, StopLimitFillModel, UnifiedFillModel
+from .slippage import (
+    NullSlippageModel,
+    ConstantSlippageModel,
+    SpreadBasedSlippageModel,
+    VolumeShareSlippageModel,
+)
 from .symbols import SymbolPropertiesProvider, SymbolProperties
 from .exchange_hours import ExchangeHoursProvider
 from .shortable import ShortableProvider, StaticShortableProvider
